dashboard/requirements.txt
flask==3.1.0
flask-limiter==3.12
flask-socketio==5.5.1
paramiko==3.5.1
requests==2.32.3
python-dotenv==1.1.0
psutil>=5.9
simple-websocket>=1.0

# Optional — system tray icon
pystray>=0.19
Pillow>=10.0

# Optional — Windows service support
pywin32>=306; sys_platform == "win32"

# Optional — build standalone .exe
pyinstaller>=6.0

# Testing
pytest>=7.0
